feat: added support for QR code based discovery of inventory data. This is helpful to maintain checks and balances which audio doesnt provide.
- Audio is still an option but QR based inventory tracking would be default with higher priority.

+++ internal/models/models.go
@@ -31,10 +31,10 @@ type DataInputApiBody struct {
 
 type InventoryData struct {
 	Item                      string  `json:"item"`
-	Quantity                  float32 `json:"quantity"`
+	Quantity                  float64 `json:"quantity"`
 	Unit                      string  `json:"unit"`
-	WholesalePricePerQuantity float32 `json:"wholesale_price_per_quantity"`
-	TotalCostOfProduct        float32 `json:"total_cost_of_product"`
+	WholesalePricePerQuantity float64 `json:"wholesale_price_per_quantity"`
+	TotalCostOfProduct        float64 `json:"total_cost_of_product"`
 }
 type SalesData struct {
 	Item                   string  `json:"item"`
